Allow non-destructive updates Changing from metadata_startup_script to metadata.startup-script.  metadata_startup_script is ForceNew, so when it is used, all gcp compute instances are destroyed and rebuilt.  See https://github.com/hashicorp/terraform-provider-google/pull/1081.  With ForceNew, the nomad acl infrastructure (as well as other data) is lost.
I have tested scaling the number of clients and confirmed that this change allows add/remove of clients without destroying the cluster.  I have not tested if it is possible to scale the number of servers.

Chaning create_before_destroy as google_computer_instance does not appear to support creation of instances with the same name.